+++ @navikt/ds-react/accordion-menu/index.ts
@@ -1,3 +1,6 @@
-export { default as AccordionMenu } from "./Menu";
-export { default as AccordionMenuCollapsable } from "./Collapsable";
-export { default as AccordionMenuItem, AccordionMenuItemType } from "./Item";
+export { default as AccordionMenu } from "./AccordionMenu";
+export { default as AccordionMenuCollapsable } from "./AccordionMenuCollapsable";
+export {
+  default as AccordionMenuItem,
+  AccordionMenuItemType,
+} from "./AccordionMenuItem";
